requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
numpy==1.24.3
faiss-cpu==1.7.4
mistralai==0.0.12
python-dotenv==1.0.0
markdown==3.5.1
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
